backend/requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
python-dotenv
sse-starlette
